- Eliminados de pago: repository, service y controller, porque Pago esta en composicion con alquiler.

+++ backend/src/main/java/ProyectoFinalTienda/TiendaVideojuegos/model/entities/PenalizacionEntity.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -32,7 +33,7 @@ public class PenalizacionEntity {
 
     @Column(nullable = false)
     @Positive(message = "El monto debe ser mayor a cero")
-    private double monto;
+    private BigDecimal monto;
 
     @Column(nullable = false)
     @NotBlank(message = "El motivo no puede estar vacío")
